PHPCS-73 Refactored things from the AbstractSniff to a suppressing sniff, to inject it in "foreign" inheritence chains
